(fixup) move TaskMetadata, LoggingContext to the root package
