+++ processor/src/main/java/com/linecorp/decaton/processor/runtime/internal/DecatonTaskRetryQueueingProcessor.java
@@ -25,7 +25,7 @@ import com.linecorp.decaton.client.internal.DecatonTaskProducer;
 import com.linecorp.decaton.client.PutTaskResult;
 import com.linecorp.decaton.processor.DecatonProcessor;
 import com.linecorp.decaton.processor.ProcessingContext;
-import com.linecorp.decaton.processor.runtime.TaskMetadata;
+import com.linecorp.decaton.processor.TaskMetadata;
 import com.linecorp.decaton.processor.metrics.Metrics;
 import com.linecorp.decaton.processor.metrics.Metrics.RetryMetrics;
 import com.linecorp.decaton.processor.runtime.RetryConfig;
