+++ processor/src/main/java/com/linecorp/decaton/processor/runtime/internal/ProcessingContextImpl.java
@@ -27,10 +27,10 @@ import com.linecorp.decaton.processor.DecatonProcessor;
 import com.linecorp.decaton.processor.runtime.DecatonTask;
 import com.linecorp.decaton.processor.DeferredCompletion;
 import com.linecorp.decaton.processor.ProcessingContext;
-import com.linecorp.decaton.processor.runtime.TaskMetadata;
+import com.linecorp.decaton.processor.TaskMetadata;
 import com.linecorp.decaton.processor.tracing.TracingProvider.ProcessorTraceHandle;
 import com.linecorp.decaton.processor.tracing.TracingProvider.RecordTraceHandle;
-import com.linecorp.decaton.processor.runtime.LoggingContext;
+import com.linecorp.decaton.processor.LoggingContext;
 import com.linecorp.decaton.processor.tracing.internal.NoopTracingProvider.NoopTrace;
 
 import lombok.extern.slf4j.Slf4j;
