+++ processor/src/test/java/com/linecorp/decaton/processor/TaskMetadataTest.java
@@ -14,13 +14,13 @@
  * under the License.
  */
 
-package com.linecorp.decaton.processor.runtime;
+package com.linecorp.decaton.processor;
 
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 
-import com.linecorp.decaton.processor.runtime.TaskMetadata;
+import com.linecorp.decaton.processor.TaskMetadata;
 
 public class TaskMetadataTest {
     @Test
